+++ aws/services/gen_services.go
@@ -105,6 +105,8 @@ import (
 	s3types "github.com/aws/aws-sdk-go-v2/service/s3/types"
 	secretsmanager "github.com/aws/aws-sdk-go-v2/service/secretsmanager"
 	secretsmanagertypes "github.com/aws/aws-sdk-go-v2/service/secretsmanager/types"
+	servicediscovery "github.com/aws/aws-sdk-go-v2/service/servicediscovery"
+	servicediscoverytypes "github.com/aws/aws-sdk-go-v2/service/servicediscovery/types"
 	sesv2 "github.com/aws/aws-sdk-go-v2/service/sesv2"
 	sesv2types "github.com/aws/aws-sdk-go-v2/service/sesv2/types"
 	sfn "github.com/aws/aws-sdk-go-v2/service/sfn"
@@ -155,6 +157,7 @@ var ServiceNames = []string{
 	"kinesis",
 	"redshift",
 	"codepipeline",
+	"cloudmap",
 	"globalaccelerator",
 	"fsx",
 	"mq",
@@ -249,6 +252,8 @@ var ResourceTypes = []string{
 	"redshiftcluster",
 	"redshiftsubnetgroup",
 	"pipeline",
+	"namespace",
+	"discoveryservice",
 	"accelerator",
 	"acceleratorlistener",
 	"fsxfilesystem",
@@ -307,6 +312,7 @@ var ServicePerAPI = map[string]string{
 	"kinesis":                 "kinesis",
 	"redshift":                "redshift",
 	"codepipeline":            "codepipeline",
+	"servicediscovery":        "cloudmap",
 	"globalaccelerator":       "globalaccelerator",
 	"fsx":                     "fsx",
 	"mq":                      "mq",
@@ -402,6 +408,8 @@ var ServicePerResourceType = map[string]string{
 	"redshiftcluster":          "redshift",
 	"redshiftsubnetgroup":      "redshift",
 	"pipeline":                 "codepipeline",
+	"namespace":                "cloudmap",
+	"discoveryservice":         "cloudmap",
 	"accelerator":              "globalaccelerator",
 	"acceleratorlistener":      "globalaccelerator",
 	"fsxfilesystem":            "fsx",
@@ -505,6 +513,8 @@ var APIPerResourceType = map[string]string{
 	"redshiftcluster":          "redshift",
 	"redshiftsubnetgroup":      "redshift",
 	"pipeline":                 "codepipeline",
+	"namespace":                "servicediscovery",
+	"discoveryservice":         "servicediscovery",
 	"accelerator":              "globalaccelerator",
 	"acceleratorlistener":      "globalaccelerator",
 	"fsxfilesystem":            "fsx",
@@ -5056,6 +5066,157 @@ func (s *Codepipeline) FetchByType(ctx context.Context, t string) (cloud.GraphAP
 
 func (s *Codepipeline) IsSyncDisabled() bool {
 	return !getBool(s.config, "aws.codepipeline.sync", true)
+}
+
+type Cloudmap struct {
+	fetcher                fetch.Fetcher
+	region, profile        string
+	config                 map[string]any
+	log                    *logger.Logger
+	ServicediscoveryClient *servicediscovery.Client
+}
+
+func NewCloudmap(cfg aws.Config, profile string, extraConf map[string]any, log *logger.Logger) cloud.Service {
+	region := cfg.Region
+	servicediscoveryClient := servicediscovery.NewFromConfig(cfg)
+
+	fetchConfig := awsfetch.NewConfig(
+		servicediscoveryClient,
+	)
+	fetchConfig.Extra = extraConf
+	fetchConfig.Log = log
+
+	return &Cloudmap{
+		ServicediscoveryClient: servicediscoveryClient,
+		fetcher:                fetch.NewFetcher(awsfetch.BuildCloudmapFetchFuncs(fetchConfig)),
+		config:                 extraConf,
+		region:                 region,
+		profile:                profile,
+		log:                    log,
+	}
+}
+
+func (s *Cloudmap) Name() string {
+	return "cloudmap"
+}
+
+func (s *Cloudmap) Region() string {
+	return s.region
+}
+
+func (s *Cloudmap) Profile() string {
+	return s.profile
+}
+
+func (s *Cloudmap) ResourceTypes() []string {
+	return []string{
+		"namespace",
+		"discoveryservice",
+	}
+}
+
+func (s *Cloudmap) Fetch(ctx context.Context) (cloud.GraphAPI, error) {
+	if s.IsSyncDisabled() {
+		return graph.NewGraph(), nil
+	}
+
+	allErrors := new(fetch.Error)
+
+	gph, err := s.fetcher.Fetch(context.WithValue(ctx, "region", s.region))
+	defer s.fetcher.Reset()
+
+	for _, e := range *fetch.WrapError(err) {
+		switch ee := e.(type) {
+		case nil:
+			continue
+		default:
+			var ae smithy.APIError
+			if errors.As(ee, &ae) && ae.ErrorMessage() == accessDenied {
+				allErrors.Add(cloud.ErrFetchAccessDenied)
+			} else {
+				allErrors.Add(ee)
+			}
+		}
+	}
+
+	if err := gph.AddResource(graph.InitResource(cloud.Region, s.region)); err != nil {
+		return gph, err
+	}
+
+	snap := gph.AsRDFGraphSnaphot()
+
+	errc := make(chan error)
+	var wg sync.WaitGroup
+	if getBool(s.config, "aws.cloudmap.namespace.sync", true) {
+		list, err := s.fetcher.Get("namespace_objects")
+		if err != nil {
+			return gph, err
+		}
+		if _, ok := list.([]servicediscoverytypes.NamespaceSummary); !ok {
+			return gph, errors.New("cannot cast to '[]servicediscoverytypes.NamespaceSummary' type from fetch context")
+		}
+		for _, r := range list.([]servicediscoverytypes.NamespaceSummary) {
+			for _, fn := range addParentsFns["namespace"] {
+				wg.Add(1)
+				go func(f addParentFn, snap tstore.RDFGraph, region string, res *servicediscoverytypes.NamespaceSummary) {
+					defer wg.Done()
+					err := f(gph, snap, region, res)
+					if err != nil {
+						errc <- err
+						return
+					}
+				}(fn, snap, s.region, &r)
+			}
+		}
+	}
+	if getBool(s.config, "aws.cloudmap.discoveryservice.sync", true) {
+		list, err := s.fetcher.Get("discoveryservice_objects")
+		if err != nil {
+			return gph, err
+		}
+		if _, ok := list.([]servicediscoverytypes.ServiceSummary); !ok {
+			return gph, errors.New("cannot cast to '[]servicediscoverytypes.ServiceSummary' type from fetch context")
+		}
+		for _, r := range list.([]servicediscoverytypes.ServiceSummary) {
+			for _, fn := range addParentsFns["discoveryservice"] {
+				wg.Add(1)
+				go func(f addParentFn, snap tstore.RDFGraph, region string, res *servicediscoverytypes.ServiceSummary) {
+					defer wg.Done()
+					err := f(gph, snap, region, res)
+					if err != nil {
+						errc <- err
+						return
+					}
+				}(fn, snap, s.region, &r)
+			}
+		}
+	}
+
+	go func() {
+		wg.Wait()
+		close(errc)
+	}()
+
+	for err := range errc {
+		if err != nil {
+			allErrors.Add(err)
+		}
+	}
+
+	if allErrors.Any() {
+		return gph, allErrors
+	}
+
+	return gph, nil
+}
+
+func (s *Cloudmap) FetchByType(ctx context.Context, t string) (cloud.GraphAPI, error) {
+	defer s.fetcher.Reset()
+	return s.fetcher.FetchByType(context.WithValue(ctx, "region", s.region), t)
+}
+
+func (s *Cloudmap) IsSyncDisabled() bool {
+	return !getBool(s.config, "aws.cloudmap.sync", true)
 }
 
 type Globalaccelerator struct {