+++ aws/services/init.go
@@ -28,6 +28,7 @@ import (
 
 var (
 	AccessService, InfraService, StorageService, MessagingService, DNSService, LambdaService, MonitoringService, CdnService, CloudformationService              cloud.Service
+	CloudmapService                                                                                                                                             cloud.Service
 	EksService, DynamodbService, SecretsmanagerService, ApigatewayService, SsmService, EfsService, CloudtrailService, CloudwatchlogsService                     cloud.Service
 	ElasticacheService, EventbridgeService, StepfunctionsService, WafService, ConfigserviceService, KinesisService, RedshiftService, CodepipelineService        cloud.Service
 	CodebuildService, BeanstalkService, CodedeployService, GlueService, SesService, CognitoService, MskService, MqService, FsxService, GlobalacceleratorService cloud.Service
@@ -81,6 +82,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	MqService = NewMq(cfg, profile, extraConf, log)
 	FsxService = NewFsx(cfg, profile, extraConf, log)
 	GlobalacceleratorService = NewGlobalaccelerator(cfg, profile, extraConf, log)
+	CloudmapService = NewCloudmap(cfg, profile, extraConf, log)
 
 	cloud.ServiceRegistry[InfraService.Name()] = InfraService
 	cloud.ServiceRegistry[AccessService.Name()] = AccessService
@@ -117,6 +119,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	cloud.ServiceRegistry[MqService.Name()] = MqService
 	cloud.ServiceRegistry[FsxService.Name()] = FsxService
 	cloud.ServiceRegistry[GlobalacceleratorService.Name()] = GlobalacceleratorService
+	cloud.ServiceRegistry[CloudmapService.Name()] = CloudmapService
 
 	awsspec.CommandFactory = &awsspec.AWSFactory{
 		Log: log,